cambio la interpretacion de las señales por las macros de waitpid y las meto en una funcion

+++ src/redirect.c
@@ -6,11 +6,20 @@
 /*   By: Dugonzal <dugonzal@student.42urduliz.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 13:05:32 by Dugonzal          #+#    #+#             */
-/*   Updated: 2023/06/11 23:42:56 by Dugonzal         ###   ########.fr       */
+/*   Updated: 2023/06/12 12:26:04 by Dugonzal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/minishell.h"
+
+void err_open(int fd)
+{
+	if (fd < 0)
+	{
+		unlink("heredoc");
+		return ;
+	}
+}
 
 void	heredoc(char *intfile, t_cmd *cmd)
 {
@@ -18,9 +27,11 @@ void	heredoc(char *intfile, t_cmd *cmd)
 	char	*line;
 
 	fd = ft_open("heredoc", 1);
+	if (fd < 0)
+	  return (err_open(fd));
 	while (42)
 	{
-		line = readline(GREEN"🙂🚀❯   "RESET);
+		line = readline(GREEN"🙂🚀❯ "RESET);
 		if (!line)
 		{
 			ft_printf (RED"\nError: heredoc readline\n"RESET);
@@ -32,10 +43,12 @@ void	heredoc(char *intfile, t_cmd *cmd)
 	}
 	close(fd);
 	cmd->intfile = ft_open("heredoc", 0);
+	if (cmd->intfile < 0)
+	  return (err_open(fd));
 	unlink("heredoc");
 }
 
-int	redir_in(t_cmd *cmd, char *intfile)
+int	redir_in_util(t_cmd *cmd, char *intfile)
 {
 	cmd->intfile = ft_open(intfile, 0);
 	if (cmd->intfile < 0)	
@@ -44,24 +57,8 @@ int	redir_in(t_cmd *cmd, char *intfile)
 		return (1);
 	return (0);
 }
-/*
 
-		if (cmd->cmd[i][0] == '<' && !cmd->cmd[i][1])
-		{
-			if (redir_in(cmd, cmd->cmd[i + 1]))
-				return (1);
-			cmd->cmd[i] = NULL;
-			cmd->cmd[i + 1] = NULL;
-			if (cmd->cmd[i + 2] && (cmd->cmd[i + 2][0] != '<'))
-			{
-				cmd->cmd[i] = cmd->cmd[i + 2];
-				cmd->cmd[i + 2] = NULL;
-			}
-			
-		}
-*/
-
-int	redir(t_cmd *cmd)
+int	redirect_input(t_cmd *cmd)
 {
 	int		i;
 	int		j;
@@ -77,7 +74,7 @@ int	redir(t_cmd *cmd)
 			if (cmd->cmd[i][j] == '<' && ft_strlen(cmd->cmd[i]) == 1)
 			{
 				cmd->cmd[i] = NULL;
-				if (redir_in(cmd, cmd->cmd[i + 1]))
+				if (redir_in_util(cmd, cmd->cmd[i + 1]))
 					return (1);
 				cmd->cmd[i + 1] = NULL;
 				if (cmd->cmd[i + 2]) //&& (cmd->cmd[i + 2][0] != '<'))
